pyproject.toml
[project]
name = "project02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
]
